+++ es/CourseManager/components/CreateLesson.js
@@ -8,6 +8,8 @@ class CreateLesson extends Component {
         this.state = {
             name: ''
         };
+
+        this.textInput = React.createRef();
     }
 
     onChangeName(event) {
@@ -24,6 +26,7 @@ class CreateLesson extends Component {
         }).then(response => {
             this.setState({ name: '' });
             this.props.addLesson(this.props.moduleIndex, response.data.data);
+            this.textInput.current.focus();
         })
         .catch(error => {
             this.props.alertError(error);
@@ -32,7 +35,7 @@ class CreateLesson extends Component {
 
     render() {
         return (
-            <form onSubmit={this.onSubmit.bind(this)}>
+            <form onSubmit={this.onSubmit.bind(this)} className="form-inline-wide">
                 <input
                     type="text"
                     className="form-control mr-1"
@@ -41,6 +44,7 @@ class CreateLesson extends Component {
                     maxLength={128}
                     value={this.state.name}
                     onChange={this.onChangeName.bind(this)}
+                    ref={this.textInput}
                 />{" "}
                 <button className="btn btn-primary">Create lesson</button>
             </form>