+++ changelog.md
@@ -5,6 +5,7 @@ Changelog
 ---
 * Use email on homepage if name is not available
 * Add name to setup screen
+* Show "no rows" message when there is no data
 
 1.3
 ---